apps/s1-filings/requirements.txt
aws-cdk-lib>=2.165.0,<3.0.0
constructs>=10.0.0,<11.0.0
boto3>=1.35.50,<1.50.0
botocore>=1.35.50,<1.50.0
types-setuptools>=75.2.0,<76.0.0
pandas>=2.2.0,<3.0.0
pyarrow>=18.0.0,<19.0.0
lxml>=5.3.0,<6.0.0
requests>=2.32.0,<3.0.0
feedparser>=6.0.11,<7.0.0